Use a buffered channel instead of spawing
The fire() func is spawned only once and will loop over a buffered
channel. The channel size can be set using graylog.BufSize, but the
default 1024 should be more than enough.

Remember the call to Fire will start to block if the channel is full
(airbrake doesn't respond).
Please not that tobi/airbrake-go *never* times out on sending the error
to airbrake, so I think this is a timebomb.
The official github.com/airbrake/gobrake has a default timeout of 10s,
which is a lot better I think.