Add option `ForceTTY` to override normal `SYS_IOCTL` detection
This adds an option to the `TextFormatter` called `ForceTTY` that
overrides the standard detection of a TTY.

The motivation for this is to be able to get logrus' pretty-printed text
output from within a tool like [Foreman][foreman] or [Forego][forego].
These tools did identify themselves as TTYs at some point during their
earlier stages, but have since stopped because it caused some side
effects that were undesirable. There is some more documentation on that
on the [Foreman wiki][foreman-wiki].

This obviously isn't a perfect solution, but does seem to accomplish
what I want. It also seems to fall in-line with some of the conventions
that are already established by other options within `TextFormatter`.
That said, I'd welcome alternatives if anyone has them. Thanks!

[forego]: https://github.com/ddollar/forego
[foreman]: https://github.com/ddollar/foreman
[foreman-wiki]: https://github.com/ddollar/foreman/wiki/Missing-Output