Golf Support
This patch adds support for the Golf package, a library that enables
types to define the fields they want returned about themselves or by way
of reflection.

For example, take the following type:

    type Person struct {
        Name  string
        Alias string
    }

    p := &Person{"Bruce", "Batman"}

If p were given to Logrus as a field value with an associated key of
"hero" then it would be formatted as such (given a text formatter):

    hero="&{Bruce Batman}"

However, the Person type may know how it wants to be formatted when it
comes to be logged as a field value, and this interface will provide
that advantage.

    func (p *Person) PlayGolf() bool {
        return true
    }

Now when a Person instance is provided to Logrus as a field value with
an associated key of hero it would be formatted as such (given a text
formatter):

    hero.Name=Bruce hero.Alias=Batman...

For more information on how the Golf package works, please see
https://github.com/akutz/golf. Thank you!