Move airbrake to dedicated repo(s)
The current implementation is using an old implementation of the
Airbrake v2 api. Airbrake has release v2 since, but this package is
necessary for in-house solutions like errbit.
The number of errbit users don't justify to continue pushing this
version as the official.
That's why the official hook is now using gobrake (the official
package), and the one coming from logrus is now named "legacy".