Add log.Logger's Output method to Logger&Entry
The standard library's *log.Logger type has the method

	Output(calldepth int, s string) error

Because it is so minimal, many libraries chose to target
it as an interface, which users could implement to receive
additional log messages:

	type Logger {
		Output(int, string) error
	}

While it is not overwhelmingly used, it is out there. Implementing
this interface allows us to use logrus.Logger and logrus.Entry
types to receive log messages from such libraries. I chose to log
such messages at the "Info" level, copying the Print* methods. The
"calldepth" argument is ignored.