Do not assume no color on windows.
recent windows 10 can support colour (it is even enabled in logrus)
however the default formatter does not use colour and so the default log
format is different on windows to other OSes when logging to the
terminal.  This removed the hard code no colour for windows terminals to
be consistent with other OSes

Now that windows coloring behaves correctly the unit test for
text_formatter was incorrect.

changes the expected test output to be the same regardless of the OS.

Co-authored-by: James Nord <jnord+github@cloudbees.com>
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>