Fix output coloring for terminals on windows
Recent versions of Windows support ANSI VT100 escape codes to e.g. allow
changing color of output. This support must be opted into by a call to
SetConsoleMode with ENABLE_VIRTUAL_TERMINAL_PROCESSING. Logrus already
has code to make this call (via the sequences package in
terminal_check_windows.go). However, color was never used due to the
check for Windows in text_formatter.go. This check was added due to the
possibility of running on old versions of Windows which don't support
this new console mode.

This change resolves these issues by attempting to set the console mode,
if it fails, we return false from checkIfTerminal, which will cause
colored output to not be used. This should allow color to work properly
on new versions of Windows, without affecting behavior on older
versions.

Signed-off-by: Kevin Parsons <kevpar@microsoft.com>