Allow filtering of fields while formatting.
One might want to filter out some fields based on some criteria. One
such criteria might be to not include fields with empty values.

Defined a FieldFilter in Formatter.go to be a function that accepts
a key (string) and a value (interface{}) and returns boolean indicating
whether to allow the field or not.
Return true if field should be allowed.
Return false if field should be filtered out.

Retrofitted TextFormatter and JSONFormatter to compose []FieldFilter and
use them (if any) while formatting.

Included a FieldFilter called IgnoreEmtpyValueFieldFilter that can be
used to filter out fields that have an empty value (zero value).